Disable `PhysicalDeviceGroupProperties` again...
I tried for a while to figure out how to support this.  The fundamental issue is that `fixed` must (by definition) be a fixed size array.  But this is an array of pointers, which varies in size by platform.  In C that's not an issue, just compile for both, in C# you can't do that.